requirements.txt
numpy==1.23.5
h5py==3.12.1
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86
pyCBD==1.2.3
pygame==2.6.1
open3d==0.19.0
matplotlib==3.10.0
pyserial==3.5
pillow~=11.2.0
scipy~=1.12.0
PyYAML~=6.0.2
setuptools~=57.4.0